Adds support for conjured items
To avoid the goblin one-shotting me, I've extended his/her/their class
instead so we could add a `conjured` attribute. A decision was made to
fix the tests as as and add test coverage in future commits, for the
sake of treating this as a short exercise that shouldn't take more than
a couple of hours.