Write tests for all types of items
Without looking at the code I've written specs based on the requirements listed and the output of the texttest.
All specs are green except for the expected 2 failing 'quality' tests for the new 'Conjured' item.
With all models tested I feel more comfortable starting the refactor. The plan at the moment is to extract the code of 1 type of item at the time into it's own method.