Reuse some methods make methods more verbose
I was doubting if instead of decreasing the quality in 2 steps, adding a parameter to the decrease_quality method and decrease it in 1 go was a better option.
I didnt go for it in the end as I like the verbosity at the moment and it's easier to follow later on why we decrease multiple times instead of e.g. decrease_quality(n) method